Created separate utils directory
