+++ components/ui/spinner.tsx
@@ -1,6 +1,6 @@
 import { Loader2Icon } from "lucide-react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/utils/utils"
 
 function Spinner({ className, ...props }: React.ComponentProps<"svg">) {
   return (
