+++ lib/db/users.service.ts
@@ -1,7 +1,7 @@
 import { getAdminFirestore } from "@/lib/firebase/firebase-admin";
 import { UserRole, UserModel } from "@/lib/auth/auth.types";
 import { Timestamp } from "firebase-admin/firestore";
-import { normalizeNotifications } from "./notifications.utils";
+import { normalizeNotifications } from "../utils/notifications.utils";
 
 const USERS_COLLECTION = "users";
